rename package to qurator.eynollah
