+++ qurator/eynollah/utils/resize.py
