+++ qurator/eynollah/writer.py
