+++ tests/test_smoke.py
@@ -1,7 +1,7 @@
 def test_utils_import():
-    import sbb_newspapers_org_image.utils
-    import sbb_newspapers_org_image.utils.contour
-    import sbb_newspapers_org_image.utils.drop_capitals
-    import sbb_newspapers_org_image.utils.drop_capitals
-    import sbb_newspapers_org_image.utils.is_nan
-    import sbb_newspapers_org_image.utils.rotate
+    import qurator.eynollah.utils
+    import qurator.eynollah.utils.contour
+    import qurator.eynollah.utils.drop_capitals
+    import qurator.eynollah.utils.drop_capitals
+    import qurator.eynollah.utils.is_nan
+    import qurator.eynollah.utils.rotate
